ProdutoControllerTest ok e pequena alteracao na controller de produto

+++ cat/src/main/java/app/Controller/ProdutoController.java
@@ -53,7 +53,7 @@ public class ProdutoController {
 	    }
 	 
 	 @GetMapping("/findById/{id}")
-	 public ResponseEntity<Produto> findById(@PathVariable int id) {
+	 public ResponseEntity<Produto> findById(@PathVariable Long id) {
 	     try {
 	    	 Produto  funcionario = this.produtoService.findById(id);
 	         return new ResponseEntity<>(funcionario, HttpStatus.OK);
